requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
beautifulsoup4==4.14.3
billiard==4.2.4
cbor2==5.7.1
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
distlib==0.4.0
Django==4.2
djangorestframework==3.16.1
exceptiongroup==1.3.1
filelock==3.19.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
kombu==5.6.1
msgpack==1.1.2
packaging==25.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
pyOpenSSL==25.3.0
pytest==9.0.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
redis==7.1.0
requests==2.32.5
service-identity==24.2.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.0
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.14
zope.interface==8.1.1
